solved

+++ divide_and_conquer/BOJ1629.c
@@ -2,8 +2,8 @@
 
 typedef long long ll;
 
-ll combine(ll val, int a, int b, int c);
-ll divide(int a, int b, int c);
+int combine(int val, int a, int b, int c);
+int divide(int a, int b, int c);
 
 int main() {
     int A, B, C;
@@ -14,15 +14,15 @@ int main() {
     return 0;
 }
 
-ll combine(ll val, int a, int b, int c) {
-    return val * val % c * ((1 + (a - 1) * (b % 2)) % c) % c;
+int combine(int val, int a, int b, int c) {
+    return (int) ((ll) val * val % c * ((1 + (a - 1) * (b % 2)) % c) % c);
 }
 
-ll divide(int a, int b, int c) {
+int divide(int a, int b, int c) {
     if (b == 1)
         return a % c;
 
-    ll val;
+    int val;
     
     val = !(b % 2) ? divide(a, b / 2, c) : divide(a, (b - 1) / 2, c);
